1.1.0

+++ web/containers/ConfigItemManager/table.tsx
@@ -1,5 +1,5 @@
 // @flow
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useContext } from 'react';
 import compose from 'recompose/compose';
 import classnames from 'classnames';
 import onlyUpdateForKeys from 'recompose/onlyUpdateForKeys';
@@ -23,6 +23,7 @@ import ReactMarkdown from 'react-markdown';
 import { maybeParseYaml, getTypeFromValue } from '../../helpers/validations';
 import { isNull, isUndefined } from 'util';
 import useMount from 'react-use/lib/useMount';
+import { InitialContext } from '../../context/init';
 
 type ConfigItemsTableProps = {
     items: Object;
@@ -36,6 +37,7 @@ type ConfigItemsTableProps = {
     levelType: string;
     stepId?: number;
     type: string;
+    definitionsOnly?: boolean;
 };
 
 const ConfigItemsTable: Function = (props: ConfigItemsTableProps) => (
@@ -150,7 +152,10 @@ let ItemsTable: Function = ({
     onDeleteStructureClick,
     t,
     type,
+    definitionsOnly,
 }: ConfigItemsTableProps) => {
+    const initContext = useContext(InitialContext);
+
     return (
         <React.Fragment>
             <Table striped condensed fixed hover>
@@ -184,7 +189,7 @@ let ItemsTable: Function = ({
                         <Th className="name text" iconName="application">
                             {t('Name')}
                         </Th>
-                        <ActionColumnHeader icon="edit" />
+                        {!definitionsOnly && <ActionColumnHeader icon="edit" />}
                         <Th className="text" iconName="info-sign">
                             {t('Value')}
                         </Th>
@@ -195,7 +200,11 @@ let ItemsTable: Function = ({
                         <Th iconName="edit">{t('Structure')}</Th>
                     </FixedRow>
                 </Thead>
-                <DataOrEmptyTable condition={!configItemsData || configItemsData.length === 0} cols={7} small>
+                <DataOrEmptyTable
+                    condition={!configItemsData || configItemsData.length === 0}
+                    cols={definitionsOnly ? (groupName ? 6 : 7) : groupName ? 7 : 8}
+                    small
+                >
                     {props => (
                         <Tbody {...props}>
                             {configItemsData.map((item: any, index: number) => (
@@ -208,44 +217,52 @@ let ItemsTable: Function = ({
                                         })}
                                     >
                                         <Td className="name">{item.name}</Td>
-                                        <ActionColumn>
-                                            <ButtonGroup>
-                                                <Button
-                                                    small
-                                                    icon="edit"
-                                                    title={t('button.edit-this-value')}
-                                                    onClick={() => {
-                                                        handleModalToggle(
-                                                            { ...item },
-                                                            (name, value, parent, isTemplatedString) => {
-                                                                onSubmit(name, value, parent, type, isTemplatedString);
-                                                                handleModalToggle(null);
-                                                            },
-                                                            intrf,
-                                                            levelType
-                                                        );
-                                                    }}
-                                                />
-                                                <Button
-                                                    small
-                                                    icon="cross"
-                                                    title={t('button.remove-this-value')}
-                                                    disabled={
-                                                        item.level ? !item.level.startsWith(levelType || '') : true
-                                                    }
-                                                    onClick={() => {
-                                                        onSubmit(
-                                                            item.name,
-                                                            null,
-                                                            item.parent_class,
-                                                            type,
-                                                            item.is_templated_string,
-                                                            true
-                                                        );
-                                                    }}
-                                                />
-                                            </ButtonGroup>
-                                        </ActionColumn>
+                                        {!definitionsOnly && (
+                                            <ActionColumn>
+                                                <ButtonGroup>
+                                                    <Button
+                                                        small
+                                                        icon="edit"
+                                                        title={t('button.edit-this-value')}
+                                                        onClick={() => {
+                                                            handleModalToggle(
+                                                                { ...item },
+                                                                (name, value, parent, isTemplatedString) => {
+                                                                    onSubmit(
+                                                                        name,
+                                                                        value,
+                                                                        parent,
+                                                                        type,
+                                                                        isTemplatedString
+                                                                    );
+                                                                    handleModalToggle(null);
+                                                                },
+                                                                intrf,
+                                                                levelType
+                                                            );
+                                                        }}
+                                                    />
+                                                    <Button
+                                                        small
+                                                        icon="cross"
+                                                        title={t('button.remove-this-value')}
+                                                        disabled={
+                                                            item.level ? !item.level.startsWith(levelType || '') : true
+                                                        }
+                                                        onClick={() => {
+                                                            onSubmit(
+                                                                item.name,
+                                                                null,
+                                                                item.parent_class,
+                                                                type,
+                                                                item.is_templated_string,
+                                                                true
+                                                            );
+                                                        }}
+                                                    />
+                                                </ButtonGroup>
+                                            </ActionColumn>
+                                        )}
                                         <Td
                                             className={`text ${item.level === 'workflow' || item.level === 'global'}`}
                                             style={{ position: 'relative' }}
@@ -277,7 +294,9 @@ let ItemsTable: Function = ({
                                                         intent="danger"
                                                         icon="trash"
                                                         onClick={() => {
-                                                            onDeleteStructureClick(item.name);
+                                                            initContext.confirmAction('ConfirmRemoveConfigItem', () => {
+                                                                onDeleteStructureClick(item.name);
+                                                            });
                                                         }}
                                                     />
                                                 )}
@@ -286,7 +305,10 @@ let ItemsTable: Function = ({
                                     </Tr>
                                     {showDescription && (
                                         <Tr>
-                                            <Td className="text" colspan={groupName ? 7 : 8}>
+                                            <Td
+                                                className="text"
+                                                colspan={definitionsOnly ? (groupName ? 6 : 7) : groupName ? 7 : 8}
+                                            >
                                                 <ReactMarkdown source={item.description} />
                                             </Td>
                                         </Tr>
